2009-04-29    Sebastian Huber <sebastian.huber@embedded-brains.de>
        * sapi/include/confdefs.h, libblock/include/rtems/bdbuf.h: Changed
        type of rtems_bdbuf_pool_configuration_size to size_t.

        * libblock/include/rtems/bdbuf.h, libblock/include/rtems/blkdev.h,
        libblock/include/rtems/diskdevs.h, libblock/src/bdbuf.c,
        libblock/src/blkdev.c, libblock/src/diskdevs.c: Buffer pool
        allocation is now cache aligned.  The cache functions are
        currently not available on all platforms so the cache line size is
        fixed to 32 bytes for now.  Changed various integer types which
        refer to block sizes, numbers and indexes.  Fixed logical block
        indexes in buffer get and read function.  It is now possible to
        delete logical disks.  Modified documentation

+++ cpukit/sapi/include/confdefs.h
@@ -740,7 +740,7 @@ rtems_fs_init_functions_t    rtems_fs_init_helper =
       rtems_bdbuf_pool_config rtems_bdbuf_pool_configuration[] = {
         {CONFIGURE_BDBUF_BUFFER_SIZE, CONFIGURE_BDBUF_BUFFER_COUNT, NULL}
       };
-      int rtems_bdbuf_pool_configuration_size = 
+      size_t rtems_bdbuf_pool_configuration_size = 
         (sizeof(rtems_bdbuf_pool_configuration) /
          sizeof(rtems_bdbuf_pool_configuration[0]));
     #endif /* CONFIGURE_INIT */
